# Performance improvements for PointVector, RayVector, SubObjectIterator