Add missing eltype methods and add type assertions to their getindex methods so that iteration over objects of these types has a chance of being typestable.

Also ensure that the return value of `size` is recognized as Tuple{Int}.

Silly microbenchmark, before:

    julia> v = PointVector{ZZRingElem}(matrix(ZZ, [1 2 3 ; ]));

    julia> @btime prod(v)
      1.217 μs (21 allocations: 336 bytes)

After:

    julia> @btime prod(v);
      288.449 ns (5 allocations: 80 bytes)