@fingolfin 2024-02-16
Timings with Julia 1.10.1 on an AMD EPYC 9554 64-Core Processor (which, despite the high core count, is currently the fastest machine I have access to, in terms of single-core performance):

Before this PR:
```
julia> @btime f(v);
  293.838 ns (11 allocations: 224 bytes)

julia> @btime prod(v);
  713.264 ns (21 allocations: 336 bytes)
```

With this PR:
```
julia> @btime f(v);
  139.726 ns (5 allocations: 80 bytes)

julia> @btime prod(v);
  139.512 ns (5 allocations: 80 bytes)
```