@fingolfin 2024-02-16
Yeah I agree - usually my MacBook is twice as fast as the older servers, and only the AMD beat it, but not by much (it doesn't clock as high as yours, of course). But in this case the old servers are as fast or faster.

For fun, here are 8 machines (including the M1 munk in the lower right) running on master (with this PR here merged) with Julia 1.10.1:

<img width="2032" alt="Screen Shot 2024-02-16 at 14 29 24" src="https://github.com/oscar-system/Oscar.jl/assets/241512/266d3ef9-b0a8-4eae-98bb-65b9565a5614">

The top left is one at home and it is on an ancient i7-3820 and it still blows away the M1. Wow.